refactoring and added some commit

+++ Ping_Pong/Ping_Pong/Boll.cpp
@@ -1,7 +1,8 @@
 #include <algorithm>
 #include "boll.h"
 
-Boll::Boll(float x, float y, float w, float h, int32_t color) : Obj(x, y, w, h, color)
+Boll::Boll(float x, float y, float w, float h, int32_t color)
+	: Obj(static_cast<uint16_t>(x), static_cast<uint16_t>(y), static_cast<uint16_t>(w), static_cast<uint16_t>(h), color)
 {
 	origin.x = static_cast<float>(RAND_MAX / 2 - rand());
 	origin.y = static_cast<float>(RAND_MAX / 2 - rand());
@@ -9,13 +10,14 @@ Boll::Boll(float x, float y, float w, float h, int32_t color) : Obj(x, y, w, h, 
 	wedge(origin, MIN_WEDGE, MAX_WEDGE);
 }
 
-Boll::Boll(const SDL_Rect & ref) : Boll(ref.x, ref.y, ref.w, ref.h) {}
+Boll::Boll(const SDL_Rect & ref)
+	: Boll(static_cast<uint16_t>(ref.x), static_cast<uint16_t>(ref.y), static_cast<uint16_t>(ref.w), static_cast<uint16_t>(ref.h)) {}
 Boll::~Boll() {}
 
 void Boll::reset(int32_t max_w, int32_t max_h)
 {
-	x = max_w / 2 - 25 / 2;
-	y = max_h / 2 - 25 / 2;
+	x = static_cast<float>(max_w) / 2.0f - 25.0f / 2.0f;
+	y = static_cast<float>(max_h) / 2.0f - 25.0f / 2.0f;
 	origin.x = static_cast<float>(RAND_MAX / 2 - rand());
 	origin.y = static_cast<float>(RAND_MAX / 2 - rand());
 	origin.norm();
